+++ SourcerySwift/Sources/SwiftTemplate.swift
@@ -229,6 +229,7 @@ open class SwiftTemplate {
 
         let arguments = [
             "xcrun",
+            "--sdk", "macosx",
             "swift",
             "build",
             "-Xswiftc", "-Onone",
@@ -240,7 +241,9 @@ open class SwiftTemplate {
                                                        currentDirectoryPath: buildDir)
 
         if compilationResult.exitCode != 0 || !compilationResult.error.isEmpty {
-            throw compilationResult.output
+            throw [compilationResult.output, compilationResult.error]
+                .filter { !$0.isEmpty }
+                .joined(separator: "\n")
         }
 
         return binaryFile